+++ pmd-core/src/test/java/net/sourceforge/pmd/lang/DummyLanguageModule.java
@@ -24,6 +24,7 @@ import net.sourceforge.pmd.lang.ast.xpath.DocumentNavigator;
 import net.sourceforge.pmd.lang.rule.AbstractRuleChainVisitor;
 import net.sourceforge.pmd.lang.rule.AbstractRuleViolationFactory;
 import net.sourceforge.pmd.lang.rule.ParametricRuleViolation;
+import net.sourceforge.pmd.lang.rule.RuleChainVisitor;
 
 import net.sf.saxon.expr.XPathContext;
 import net.sf.saxon.sxpath.IndependentContext;
@@ -37,7 +38,7 @@ public class DummyLanguageModule extends BaseLanguageModule {
     public static final String TERSE_NAME = "dummy";
 
     public DummyLanguageModule() {
-        super(NAME, null, TERSE_NAME, DummyRuleChainVisitor.class, "dummy");
+        super(NAME, null, TERSE_NAME, "dummy");
         addVersion("1.0", new Handler());
         addVersion("1.1", new Handler());
         addVersion("1.2", new Handler());
@@ -49,6 +50,11 @@ public class DummyLanguageModule extends BaseLanguageModule {
         addVersion("1.8", new Handler(), "8");
     }
 
+    /**
+     * @deprecated for removal with PMD 7. A language dependent rule chain visitor is not needed anymore.
+     *      See {@link RuleChainVisitor}.
+     */
+    @Deprecated
     public static class DummyRuleChainVisitor extends AbstractRuleChainVisitor {
         @Override
         protected void visit(Rule rule, Node node, RuleContext ctx) {